# docs/swagger.yaml
basePath: /
definitions:
  dto.AdminLogRespDTO:
    properties:
      action:
        type: string
      details:
        type: string
      id:
        type: integer
      performed_at:
        type: string
      user:
        $ref: '#/definitions/dto.UserResponseDTO'
    type: object
...
  dto.RoleDTO:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      type:
        $ref: '#/definitions/model.RoleType'
      updated_at:
        type: string
      users:
        items:
          $ref: '#/definitions/dto.UserResponseDTO'
        type: array
    type: object
...
  dto.UserResponseDTO:
    properties:
      avatar_file_name:
        type: string
      created_at:
        type: string
      created_by:
        $ref: '#/definitions/dto.UserResponseDTO'
      created_by_id:
        type: integer
      deleted_by_id:
        type: integer
      display_name:
        type: string
      email:
        type: string
      id:
        type: integer
      role:
        $ref: '#/definitions/dto.RoleDTO'
      role_id:
        type: integer
      status:
        $ref: '#/definitions/model.UserStatusType'
      updated_at:
        type: string
      updated_by:
        $ref: '#/definitions/dto.UserResponseDTO'
      updated_by_id:
        type: integer
      username:
        type: string
    type: object
...
  model.RoleType:
    enum:
    - super_admin
    - admin
    - streamer
    - user
    type: string
    x-enum-varnames:
    - SUPPERADMINROLE
    - ADMINROLE
    - STREAMER
    - USERROLE
  model.UserStatusType:
    enum:
    - online
    - offline
    - blocked
    type: string
    x-enum-varnames:
    - ONLINE
    - OFFLINE
    - BLOCKED
  utils.PaginationModel-dto_AdminLogRespDTO:
    properties:
      current_page:
        type: integer
      exec_time:
        type: number
      index:
        type: integer
      is_new_filter:
        type: boolean
      length:
        type: integer
      next:
        type: integer
      obj:
        additionalProperties: true
        type: object
      page:
        items:
          $ref: '#/definitions/dto.AdminLogRespDTO'
        type: array
      page_size:
        type: integer
      previous:
        type: integer
      query:
        type: string
      route:
        type: string
      total_items:
        type: integer
    type: object
...
host: localhost:8686
info:
  contact: {}
  description: Swagger API Admin Live Stream.
  title: Admin API Live Stream
  version: "1.0"
paths:
  /api/admins/{id}:
    get:
      consumes:
      - application/json
      description: Get admin details by ID
      parameters:
      - description: Admin ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Admin details
          schema:
            $ref: '#/definitions/dto.UserResponseDTO'
        "400":
          description: Invalid ID parameter or not found
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: Get Admin by ID
      tags:
      - Admin
  /api/admins/logs:
    get:
      consumes:
      - application/json
      description: Get logs for the current admin
      parameters:
      - enum:
        - action
        - details
        - id
        - username
        - email
        in: query
        name: filter_by
        type: string
      - in: query
        name: is_me
        type: boolean
      - in: query
        maxLength: 255
        name: keyword
        type: string
      - in: query
        maximum: 20
        minimum: 1
        name: limit
        type: integer
      - in: query
        minimum: 1
        name: page
        type: integer
      - enum:
        - DESC
        - ASC
        in: query
        name: sort
        type: string
      - enum:
        - performed_at
        in: query
        name: sort_by
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Admin logs
          schema:
            $ref: '#/definitions/utils.PaginationModel-dto_AdminLogRespDTO'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: Get Admin Logs
      tags:
      - Admin
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"